requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
Flask==2.3.3
aiohttp==3.8.5
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
gunicorn==21.2.0
